+++ src/main/scala/org/skycastle/server/game/shape/Block.scala
@@ -6,6 +6,6 @@ import _root_.org.skycastle.server.game.components.Shape
  * 
  */
 
-class Block(width: Double, height: Double, depth: Double) extends Shape {
+class Block(width: Float, height: Float, depth: Float) extends Shape {
   
 }
