+++ src/main/scala/org/skycastle/server/game/shape/Cloud.scala
@@ -6,6 +6,6 @@ import _root_.org.skycastle.server.game.components.Shape
  * 
  */
 
-class Cloud(diameter: Double) extends Shape {
+class Cloud(diameter: Float) extends Shape {
   
 }
