+++ src/main/scala/org/skycastle/server/game/shape/Lump.scala
@@ -5,7 +5,7 @@ import _root_.org.skycastle.server.game.components.Shape
 /**
  * Undefined lump, e.g. stone.
  */
-class Lump(diameter: Double) extends Shape{
+class Lump(diameter: Float) extends Shape{
   
 }
