+++ src/main/scala/org/skycastle/server/game/shape/Pile.scala
@@ -6,7 +6,7 @@ import _root_.org.skycastle.server.game.components.Shape
  * 
  */
 
-class Pile(baseDiameter: Double, height: Double) extends Shape{
+class Pile(baseDiameter: Float, height: Float) extends Shape{
   
 }
