+++ src/main/scala/org/skycastle/server/game/shape/Rod.scala
@@ -5,7 +5,7 @@ import _root_.org.skycastle.server.game.components.Shape
 /**
  * 
  */
-class Rod(diameter: Double, length: Double) extends Shape{
+class Rod(diameter: Float, length: Float) extends Shape{
   
 }
