+++ src/main/scala/org/skycastle/server/game/shape/Thread.scala
@@ -6,7 +6,7 @@ import _root_.org.skycastle.server.game.components.Shape
  * 
  */
 
-class Thread(diameter: Double, length: Double) extends Shape {
+class Thread(diameter: Float, length: Float) extends Shape {
   
 }
